feat: to f32

+++ deno/tree.test.ts
@@ -5,13 +5,13 @@ await init();
 Deno.test({
   name: "tree",
   fn: async () => {
-    expect(findInside(new Float64Array([0, 0, 1, 1, 2, 2]), [1, 1])).toBe(
+    expect(findInside(new Float32Array([0, 0, 1, 1, 2, 2]), [1, 1])).toBe(
       1,
     );
-    expect(findInside(new Float64Array([0, 0, 1, 1, 2, 2]), [0, 0])).toBe(
+    expect(findInside(new Float32Array([0, 0, 1, 1, 2, 2]), [0, 0])).toBe(
       0,
     );
-    expect(findInside(new Float64Array([0, 0, 1, 1, 2, 2]), [1.9, 1.8]))
+    expect(findInside(new Float32Array([0, 0, 1, 1, 2, 2]), [1.9, 1.8]))
       .toBe(2);
   },
 });
